requirements.txt
configparser
requests
python-telegram-bot